refactor: 스왑 뷰모델 리펙토링
