+++ app/src/main/java/com/example/swapit/ui/swap/SwapViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.swapit.ui.shopping.detail.select
+package com.example.swapit.ui.swap
 
 import android.util.Log
 import androidx.compose.runtime.MutableState
@@ -12,7 +12,7 @@ import com.example.swapit.domain.repository.SwapRepository
 import com.example.swapit.ui.base.BaseViewModelFactory
 import kotlinx.coroutines.launch
 
-class SwapProductViewModel(private val repository: SwapRepository) : ViewModel() {
+class SwapViewModel(private val repository: SwapRepository) : ViewModel() {
     val requestedProductId = mutableLongStateOf(0)
     val targetProductId = mutableLongStateOf(0)
     private val _dialogStates = mutableStateOf(mutableMapOf<Long, Boolean>())
@@ -60,7 +60,7 @@ class SwapProductViewModel(private val repository: SwapRepository) : ViewModel()
 
         fun factory(repository: SwapRepository): ViewModelProvider.Factory =
             BaseViewModelFactory {
-                SwapProductViewModel(repository = repository)
+                SwapViewModel(repository = repository)
             }
     }
 }
